Refactor to org.protelis
