+++ protelis.parser.tests/META-INF/MANIFEST.MF
@@ -21,4 +21,4 @@ Import-Package: org.apache.log4j,
  org.junit.runners.model;version="4.5.0",
  org.hamcrest.core
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
-Export-Package: it.unibo.alchemist.language.protelis
+Export-Package: org.protelis.parser
