--- protelis.parser.ui/src/it/unibo/alchemist/language/protelis/ui/ProtelisUiModule.java
@@ -1,15 +1,0 @@
-/*
- * generated by Xtext
- */
-package it.unibo.alchemist.language.protelis.ui;
-
-import org.eclipse.ui.plugin.AbstractUIPlugin;
-
-/**
- * Use this class to register components to be used within the IDE.
- */
-public class ProtelisUiModule extends it.unibo.alchemist.language.protelis.ui.AbstractProtelisUiModule {
-	public ProtelisUiModule(AbstractUIPlugin plugin) {
-		super(plugin);
-	}
-}
