--- protelis.parser.ui/src/it/unibo/alchemist/language/protelis/ui/contentassist/ProtelisProposalProvider.xtend
@@ -1,12 +1,0 @@
-/*
- * generated by Xtext
- */
-package it.unibo.alchemist.language.protelis.ui.contentassist
-
-import it.unibo.alchemist.language.protelis.ui.contentassist.AbstractProtelisProposalProvider
-
-/**
- * see http://www.eclipse.org/Xtext/documentation.html#contentAssist on how to customize content assistant
- */
-class ProtelisProposalProvider extends AbstractProtelisProposalProvider {
-}
