--- protelis.parser.ui/src/it/unibo/alchemist/language/protelis/ui/labeling/ProtelisDescriptionLabelProvider.xtend
@@ -1,24 +1,0 @@
-/*
-* generated by Xtext
-*/
-package it.unibo.alchemist.language.protelis.ui.labeling
-
-//import org.eclipse.xtext.resource.IEObjectDescription
-
-/**
- * Provides labels for a IEObjectDescriptions and IResourceDescriptions.
- * 
- * see http://www.eclipse.org/Xtext/documentation.html#labelProvider
- */
-class ProtelisDescriptionLabelProvider extends org.eclipse.xtext.ui.label.DefaultDescriptionLabelProvider {
-
-	// Labels and icons can be computed like this:
-	
-//	override text(IEObjectDescription ele) {
-//		ele.name.toString
-//	}
-//	 
-//	override image(IEObjectDescription ele) {
-//		ele.EClass.name + '.gif'
-//	}	 
-}
