--- protelis.parser.ui/src/it/unibo/alchemist/language/protelis/ui/labeling/ProtelisLabelProvider.xtend
@@ -1,29 +1,0 @@
-/*
-* generated by Xtext
-*/
-package it.unibo.alchemist.language.protelis.ui.labeling
-
-import com.google.inject.Inject
-
-/**
- * Provides labels for a EObjects.
- * 
- * see http://www.eclipse.org/Xtext/documentation.html#labelProvider
- */
-class ProtelisLabelProvider extends org.eclipse.xtext.ui.label.DefaultEObjectLabelProvider {
-
-	@Inject
-	new(org.eclipse.emf.edit.ui.provider.AdapterFactoryLabelProvider delegate) {
-		super(delegate);
-	}
-
-	// Labels and icons can be computed like this:
-	
-//	def text(Greeting ele) {
-//		'A greeting to ' + ele.name
-//	}
-//
-//	def image(Greeting ele) {
-//		'Greeting.gif'
-//	}
-}
