--- protelis.parser.ui/src/it/unibo/alchemist/language/protelis/ui/outline/ProtelisOutlineTreeProvider.xtend
@@ -1,13 +1,0 @@
-/*
-* generated by Xtext
-*/
-package it.unibo.alchemist.language.protelis.ui.outline
-
-/**
- * Customization of the default outline structure.
- *
- * see http://www.eclipse.org/Xtext/documentation.html#outline
- */
-class ProtelisOutlineTreeProvider extends org.eclipse.xtext.ui.editor.outline.impl.DefaultOutlineTreeProvider {
-	
-}
