+++ protelis.parser.ui/src/org/protelis/parser/ui/ProtelisUiModule.java
@@ -1,0 +1,15 @@
+/*
+ * generated by Xtext
+ */
+package org.protelis.parser.ui;
+
+import org.eclipse.ui.plugin.AbstractUIPlugin;
+
+/**
+ * Use this class to register components to be used within the IDE.
+ */
+public class ProtelisUiModule extends org.protelis.parser.ui.AbstractProtelisUiModule {
+	public ProtelisUiModule(AbstractUIPlugin plugin) {
+		super(plugin);
+	}
+}
