+++ protelis.parser.ui/src/org/protelis/parser/ui/contentassist/ProtelisProposalProvider.xtend
@@ -1,0 +1,13 @@
+/*
+ * generated by Xtext
+ */
+package org.protelis.parser.ui.contentassist
+
+import org.protelis.parser.ui.contentassist.AbstractProtelisProposalProvider
+
+/**
+ * See https://www.eclipse.org/Xtext/documentation/304_ide_concepts.html#content-assist
+ * on how to customize the content assistant.
+ */
+class ProtelisProposalProvider extends AbstractProtelisProposalProvider {
+}
