+++ protelis.parser.ui/src/org/protelis/parser/ui/outline/ProtelisOutlineTreeProvider.xtend
@@ -1,0 +1,13 @@
+/*
+ * generated by Xtext
+ */
+package org.protelis.parser.ui.outline
+
+/**
+ * Customization of the default outline structure.
+ *
+ * See https://www.eclipse.org/Xtext/documentation/304_ide_concepts.html#outline
+ */
+class ProtelisOutlineTreeProvider extends org.eclipse.xtext.ui.editor.outline.impl.DefaultOutlineTreeProvider {
+	
+}
