--- protelis.parser/src/it/unibo/alchemist/language/protelis/ProtelisRuntimeModule.java
@@ -1,11 +1,0 @@
-/*
- * generated by Xtext
- */
-package it.unibo.alchemist.language.protelis;
-
-/**
- * Use this class to register components to be used at runtime / without the Equinox extension registry.
- */
-public class ProtelisRuntimeModule extends it.unibo.alchemist.language.protelis.AbstractProtelisRuntimeModule {
-
-}
