--- protelis.parser/src/it/unibo/alchemist/language/protelis/ProtelisStandaloneSetup.java
@@ -1,16 +1,0 @@
-/*
-* generated by Xtext
-*/
-package it.unibo.alchemist.language.protelis;
-
-/**
- * Initialization support for running Xtext languages 
- * without equinox extension registry
- */
-public class ProtelisStandaloneSetup extends ProtelisStandaloneSetupGenerated{
-
-	public static void doSetup() {
-		new ProtelisStandaloneSetup().createInjectorAndDoEMFRegistration();
-	}
-}
-
