--- protelis.parser/src/it/unibo/alchemist/language/protelis/formatting/ProtelisFormatter.xtend
@@ -1,30 +1,0 @@
-/*
- * generated by Xtext
- */
-package it.unibo.alchemist.language.protelis.formatting
-
-import org.eclipse.xtext.formatting.impl.AbstractDeclarativeFormatter
-import org.eclipse.xtext.formatting.impl.FormattingConfig
-// import com.google.inject.Inject;
-// import it.unibo.alchemist.language.protelis.services.ProtelisGrammarAccess
-
-/**
- * This class contains custom formatting description.
- * 
- * see : http://www.eclipse.org/Xtext/documentation.html#formatting
- * on how and when to use it 
- * 
- * Also see {@link org.eclipse.xtext.xtext.XtextFormattingTokenSerializer} as an example
- */
-class ProtelisFormatter extends AbstractDeclarativeFormatter {
-
-//	@Inject extension ProtelisGrammarAccess
-	
-	override protected void configureFormatting(FormattingConfig c) {
-// It's usually a good idea to activate the following three statements.
-// They will add and preserve newlines around comments
-//		c.setLinewrap(0, 1, 2).before(SL_COMMENTRule)
-//		c.setLinewrap(0, 1, 2).before(ML_COMMENTRule)
-//		c.setLinewrap(0, 1, 1).after(ML_COMMENTRule)
-	}
-}
