--- protelis.parser/src/it/unibo/alchemist/language/protelis/validation/ProtelisValidator.xtend
@@ -1,24 +1,0 @@
-/*
- * generated by Xtext
- */
-package it.unibo.alchemist.language.protelis.validation
-//import org.eclipse.xtext.validation.Check
-
-/**
- * Custom validation rules. 
- *
- * see http://www.eclipse.org/Xtext/documentation.html#validation
- */
-class ProtelisValidator extends AbstractProtelisValidator {
-
-//  public static val INVALID_NAME = 'invalidName'
-//
-//	@Check
-//	def checkGreetingStartsWithCapital(Greeting greeting) {
-//		if (!Character.isUpperCase(greeting.name.charAt(0))) {
-//			warning('Name should start with a capital', 
-//					MyDslPackage.Literals.GREETING__NAME,
-//					INVALID_NAME)
-//		}
-//	}
-}
