+++ protelis.parser/src/org/protelis/parser/ProtelisRuntimeModule.java
@@ -1,0 +1,11 @@
+/*
+ * generated by Xtext
+ */
+package org.protelis.parser;
+
+/**
+ * Use this class to register components to be used at runtime / without the Equinox extension registry.
+ */
+public class ProtelisRuntimeModule extends org.protelis.parser.AbstractProtelisRuntimeModule {
+
+}
