+++ protelis.parser/src/org/protelis/parser/ProtelisStandaloneSetup.java
@@ -1,0 +1,18 @@
+/*
+* generated by Xtext
+*/
+package org.protelis.parser;
+
+import org.protelis.parser.ProtelisStandaloneSetupGenerated;
+
+/**
+ * Initialization support for running Xtext languages 
+ * without equinox extension registry
+ */
+public class ProtelisStandaloneSetup extends ProtelisStandaloneSetupGenerated{
+
+	public static void doSetup() {
+		new ProtelisStandaloneSetup().createInjectorAndDoEMFRegistration();
+	}
+}
+
